ui-next: refactor code structure
