+++ packages/ui-next/api.ts
@@ -1,4 +1,4 @@
 export type { Context } from './src/context';
-export { usePageData } from './src/pageData';
-export type { PageData } from './src/pageData';
+export { usePageData } from './src/context/pageData';
+export type { PageData } from './src/context/pageData';
 export type { ComponentName } from './src/registry';
